Layout
This layout was made in the first session of HTML classthat's Full Stack Udemy. The main idea was to create a University page and I decided to create a portfolio idea.
My own ideas were emerging and still need complements, thats means  it contains mistakes that can be improved during the course.